test: move new parseArgs null-config tests to a .test.ts file
