+++ test/js/node/util/parse_args/parse-args.test.mjs
@@ -1129,51 +1129,4 @@ describe("parseArgs extra tests", () => {
     expect(parseArgs({ allowPositionals: undefined })).toEqual({ values: { __proto__: null }, positionals: [] });
     expect(parseArgs({ allowPositionals: null })).toEqual({ values: { __proto__: null }, positionals: [] });
   });
-
-  describe("null top-level boolean config flags fall back to defaults", () => {
-    test.each([null, undefined])("%p acts like the key being absent for all four flags", absent => {
-      const base = { values: { __proto__: null }, positionals: [] };
-      expect(parseArgs({ args: [], options: {}, strict: absent })).toEqual(base);
-      expect(parseArgs({ args: [], options: {}, tokens: absent })).toEqual(base);
-      expect(parseArgs({ args: [], options: {}, allowPositionals: absent })).toEqual(base);
-      expect(parseArgs({ args: [], options: {}, allowNegative: absent })).toEqual(base);
-    });
-
-    test("strict: null uses default true (unknown option still rejected)", () => {
-      expect(() => parseArgs({ args: ["--nope"], options: {}, strict: null })).toThrow(
-        expect.objectContaining({ code: "ERR_PARSE_ARGS_UNKNOWN_OPTION" }),
-      );
-    });
-
-    test("tokens: null uses default false (no tokens array in result)", () => {
-      const result = parseArgs({ args: [], options: {}, tokens: null });
-      expect(Object.hasOwn(result, "tokens")).toBe(false);
-    });
-
-    test("allowNegative: null uses default false (--no-foo not recognized)", () => {
-      const options = { foo: { type: "boolean" } };
-      expect(() => parseArgs({ args: ["--no-foo"], options, allowNegative: null })).toThrow(
-        expect.objectContaining({ code: "ERR_PARSE_ARGS_UNKNOWN_OPTION" }),
-      );
-    });
-
-    test("allowPositionals: null uses default !strict", () => {
-      expect(() => parseArgs({ args: ["pos"], options: {}, strict: null, allowPositionals: null })).toThrow(
-        expect.objectContaining({ code: "ERR_PARSE_ARGS_UNEXPECTED_POSITIONAL" }),
-      );
-      expect(parseArgs({ args: ["pos"], options: {}, strict: false, allowPositionals: null })).toEqual({
-        values: { __proto__: null },
-        positionals: ["pos"],
-      });
-    });
-
-    test.each(["strict", "tokens", "allowPositionals", "allowNegative"])(
-      "non-null non-boolean %s still throws ERR_INVALID_ARG_TYPE",
-      key => {
-        expect(() => parseArgs({ args: [], options: {}, [key]: "yes" })).toThrow(
-          expect.objectContaining({ code: "ERR_INVALID_ARG_TYPE" }),
-        );
-      },
-    );
-  });
 });
